sound_stm: Replace frequency synthesizer driver with clocking framework
This patch is the sixth in a series designed to incrementally add working ALSA
support to the SH kernel. It removes the frequency synthesizer driver so the
clocking framework is now relied upon to set the correct frequency rate. A custom
wrapper to clk_get/clk_put provides clocking adjustments via ALSA controls.

Signed-off-by: Pawel Moll <pawel.moll@st.com>
Signed-off-by: John Boddie <john.boddie@st.com>