sound_stm: Claim and release pad PIOs as required for STx7141
This patch is the eigth in a series designed to incrementally add working ALSA
support to the SH kernel. It will add code to the PCM player/reader and SPDIF
player to ensure that the pad PIOs are claimed on driver probe and released on
driver removal. Currently only STx7141 requires this.

Signed-off-by: Pawel Moll <pawel.moll@st.com>
Signed-off-by: John Boddie <john.boddie@st.com>