sata_stm: Configure miphy correctly for PCIe
Don't unconditionally flip the clock for miphy lane1 to 30MHz. If
we are using PCIe we need the 100MHz external clock that must be
present, and this is fine to talk to lane0. If it isn't present
then we have to use the 30MHz clock for lane1 to allow us to talk
to lane0

Signed-off-by: David McKay <david.mckay@st.com>