stm: stx7108: Work-around for HoM-MTD issue
This patch fixes the issue reported in bugzilla 10908.
The sequence of a HoM hibernate and resume, followed by a suspend,
causes the subsequent resume to hang the system.

After analysis it seems there is an issue in the MTD subsystem (just
removing the MTD the system works fine) but I find-out that
bypassing the EMI PM the system works fine too and Linux
is able to read/write the Flash.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>