stm_usb: Fix shutdown controller properly before poweroff.
This patch add shutdown function to both EHCI and OHCI platform drivers
to shutdown USB controller before they are powered off. Without this
patch there is a high chance of STBus hangs, because of active or
pending dma requests in stbus form USB controllers.

Originally this bug was discovered as part of Bugzilla 11831 triage.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>