stm_mali: Add workaround to mmap if can't stall Mali MMU
Make the mmap code do the same thing an unmap, timing out and continuing
regardless if an mmu doesn't go into a stall state. This prevents the
driver going into an unrecoverable busy loop, requiring a reboot.
Obviously we would still like to know why the MMUs don't repsond to the
stall commands on rare occasions

Signed-off-by: Alvar Bray <alvar.bray@st.com>