stm_mali: Add some more memory barrier paranoia
In the original git://git.bri.st.com/stm/mali400 codebase this is:

6847a741646abfceb2d4d4f87b9e5c4ed3497869 Add some more memory barrier paranoia.

This was taken from the r2p0 codebase. The original comment was:

Add some more memory barrier paranoia, doesn't look to really
effect the on-going VgMark issues, but leave it in for now
although it can be switched off.

Signed-off-by: Alvar Bray <alvar.bray@st.com>