sh: fix issue in xchg_u32, xchg_u8 and __cmpxchg_u32 functions
This is a continuation of the patch in commit
bfe2d986381f530504fe2dd80da32c55834afe1d "sh: fixed issue in xchg_u32
function."

Anyway repeating the comments:

This patch addresses a use-case when one of the input parameter to
xchg_u32 or xchg_u8 or __cmpxchg_u32 inline asm function is equal to
r15(stack-pointer).

For example:
Code in exit_mm() function passes address of structure(val) to
xchg_u32 which is basically a stack pointer(r15). xchg_u32 always sets m
to -4(0xfffffffc). Which is incorrect(Actual Bug).

Considering there could be one or more instances of this kind in the
current or future code, this patch is must required if gUSA is used.

This patch changes input constraint to an input/output constraint so that
the compiler cannot pass r15 directly and must use a temporary register
instead.

Originally this bug was discovered as part of stlinux bugzilla 11229 triage.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>