Updated defconfigs for STLinux-ASV
This patch is aimed to fix/synch/extend defconfigs for ST supported boards.
In order to test as much devices and features as possible, it is helpful to enable
all these subsystems and options in defconfigs.

Major changes consists of enabling (where needed):
* SATA/PATA
* co-processors
* Power Management (including HoM)
* LiRC
* USB
* Bluetooth
* RTC LPC
* Posix mqueue (needed by LTP)
* Optimise for size (-Os)
* Fast module loader
* Calibrate delay (sh4 specific)
* MTD

Further defconfigs have been synchronised with related Kconfigs.

Signed-off-by: Francesco Rundo <francesco.rundo@st.com>
Reviewed-by: Carmelo Amoroso <carmelo.amoroso@st.com>