stm_pad: ensure PIO mode is selected for GPIO pads
Previously it was assumed that pads requested by the GPIO framework were already
in PIO mode.  However, it turns out that this is not always true (e.g. on
stx7108c2, the SPI-Boot pads default to ALT1).  As a result, certain pads were
routed to their alternate IP blocks, and not usable by GPIO clients.

This patch updates the code that requests a pad for use by the GPIO library,
adding a call to set the function to the registered 'stm_pad_gpio_function'
number.  The patch also updates the pad initialisation on some SoCs to ensure
the specified 'stm_pad_gpio_function' number actually agrees with the SoC's
config_pio() function.

Signed-off-by: Angus Clark <angus.clark@st.com>