stm_nand_emi: update EMI timing configuration to match NAND controller
This updates the EMI timing configuration to match that assumed to be used by
the NAND controller.  This allows the same timing parameters to be used with all
the STM NAND drivers (stm_nand_emi, stm_nand_flex and stm_nand_afm).  All
supported platforms have been re-tested following this update.

Note, the configuration of the NAND timing parameters is still not fully
understood; it has been confirmed that the description in the functional spec is
not correct!  We are awaiting clarification from Design and Validation, after
which, it is likely that the NAND drivers will need to be updated once more!

Signed-off-by: Angus Clark <angus.clark@st.com>