stm_nand_afm: workaround for SPARE3 arbitration bug
This patch implements a workaround for the AFM "SPARE3" bug (see RnDHV00032925).
When executing the SPARE3 command, the NAND controller fails to relinquish the
EMI bus if it must stall due to the FIFO status (i.e. FULL on reads, or EMPTY on
writes).  If the CPU initiates an access to a different EMI bank during the
SPARE3 command (or more accurately, during the 48 bytes of "dummy" data), then
the EMI subsystem will enter deadlock; the other EMI access will stall until the
NAND FIFO is drained/filled, and the CPU will not drain/fill the FIFO until the
other EMI access has returned.

To workaround this issue, the AFM driver has been modified to not use the SPARE3
command.  For OOB read operations we use the DATA0 command to read 64 bytes from
NAND, inserting the ECC data manually from the ecc_checkcode registers.  This
has the side-affect of over-reading 48 bytes from the end of the spare area.
This is thought to be benign for current NAND devices, and has little impact on
overall performance.  For write operations, we have no other option but to
fall-back to FLEX mode.

Signed-off-by: Angus Clark <angus.clark@st.com>