stm_nand_afm: support writing bad block markers to OOB
Since the bad-block scanning code now handles boot-mode ECC and AFM ECC, it is
no longer mandatory to use Flash-resident BBTs (although still highly
recommended!).  This patch updates the way in which blocks are marked bad.  If
Flash-resident BBTs are used, then the BBTs are updated accordingly and written
back to Flash, as before.  However, if Flash-resident BBTs are not used, then
the bad-block marker is written to the OOB area.  We also 'wipe' the first AFM
marker in OOB (just in case it survived the preceding failed ERASE operation) to
prevent subsequent on-boot scans from recognising an AFM block, instead of a
marked-bad block.

Signed-off-by: Angus Clark <angus.clark@st.com>