stm_pad: fix regression in GPIO pad configuration
Previously, the STM GPIO interface assumed that PIO pins were either
already in GPIO mode, or GPIO mode could be selected purely by
configuring the PIO's "PCx" registers.  However, this assumption
breaks on some SoCs (e.g. stx7141, stx7108) where additional
configuration is required to select GPIO mode (typically via a SYS_CFG
register).  In such cases, the GPIO framework would fail to configure
GPIO mode if the pin had previously been configured in ALT mode
(e.g. SPIBoot pins on reset).

Commit dfa8fbd1471cee18b8a4eae8685004fe559b8e48 "stm_pad: ensure PIO
mode is selected for GPIO pads" attempted to fix this issue by calling
the SoC's stm_pad_config() function when requesting the pad for use
with the GPIO framework.  However, this introduced a bug on some
platforms where we require an ALT function of '-1' to be passed in
when configuring an input pin.  Note, '-1' is used in this case to
highlight the fact that such inputs are permanently connected to all
their alternate peripheral functions and GPIO function, and hence,
specifying an ALT function does not make sense.

This patch fixes the regression, while maintaining the semantics of
using '-1' where an ALT function should not be specified.  When
registering the SoC's stm_config_pad() function, we now specify two
GPIO function numbers, one to be used when configuring a GPIO input,
and one for configuring a GPIO output.

On most SoCs, the input GPIO number is '-1', indicating that the
inputs are permanently connected to all ALT peripheral and GPIO
functions.  At present, the only exception is the stx7108, where ALT
functions select routing of both input and output signals.

The output GPIO number is typically '0' (i.e. ALT0 == GPIO).  However,
on some platforms (e.g. stx5206), the documentation defines ALT
functions starting from 0.  In order to be consistent with the
documentation, we define some other function number to represent GPIO
mode.

The other part of this change is that the gpio functions to configure
the direction (stm_gpio_direction_input and stm_gpio_direction_output)
no longer write directly to the PIO hardware, but configure the pin
using the pad manager through a new private function,
stm_pad_configure_gpio(). This ensures that the SoC specific code to
configure the pad is called. For most devices this already calls
stm_gpio_direction() to set the PIO registers, and this call just
needs to be added to the 7108 pio_config function.

Equivalent changes have also been made to the deprecated stpio_
interface: __stpio_request_pin and stpio_configure_pin now call
stm_pad_configure_gpio instead of __stm_gpio_direction.

Signed-off-by: Angus Clark <angus.clark@st.com>