stm_cpufreq: Tidy up the stm_cpufreq API
As most SoCs now have the same clockgen based CPU clock generation
hardware, the CPU clock is always called "cpu_clk", and clk_set_rate()
work correctly, most of the SoC specific cpufreq code can be replaced
by a single implementation, which can support a greater range of
frequencies than the previous implementations.

The only exception to this is the STx7200, which still requires its
own code as the clock LLA code still doesn't support it fully.

This means that the stm_cpufreq API needs to change, so that each
SoC has to report an array with all the divisors supported, as it
may no longer be simple integer divisor.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>