pci_stm: Fix PCI on 7108
PCI on 7108 was broken due to the EMI_NOT_PCI bit being set incorrectly.
This bit must be set in order to actually enable the req0/req3
workaround. It is simpler to set it all the time, as the EMI mode really
means MPX mode (a cut down, limited PCI) which is not supported.

This patch also removes the unused req0_emi field, which is redundant.

Signed-off-by: David McKay <david.mckay@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>