pci_stm: Remove over-zealous BUG_ON for REQ_GNT0
It is not a bug to not use REQ_GNT0. It is a dedicated pin,
but you do not have to use it. If you leave it unconnected there
will be no problems as the code will mask it out anyway. If you do
set it to PCI_PIN_DEFAULT, it must be externally pulled up.

Signed-off-by: David McKay <david.mckay@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>