pci_stm: Remove assumption that PCI is bus 0
The code was assuming that the pci-emiss bus was always bus 0.
When a PCI express controller is also present, this may not be
the case depending on which is registered first. The code needs
to know it it talking to the root bus to decide to drive type 0
or type 1 accesses.

Signed-off-by: David McKay <david.mckay@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>