pcie_stm: Add PCI express support for 7108c2
Adds support for the PCI express cell found on the 7108 cut2. Also
includes support for MSI (Message Signalled Interrupts).

On the 7108c2 legacy interrupts are broken, the driver for a particular
card must use MSI interrupts or it will not function. Many drivers do
not use MSI by default, so some modification of the driver may be
required.

Signed-off-by: David McKay <david.mckay@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>