pci_stm: Update 7105 PCI code to use pad manager
The PCI REQ/GNT lines were not being properly configured on the 7105,
the sysconfig settings to set the correct alternative function were not
being done. The code has been updated to use the pad manager to claim
the PIOs used by the REQ/GNT lines and the interrupts. The pad manager
code will correctly configure the sysconfig registers.

Signed-off-by: David McKay <david.mckay@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>