pcie_stm: Disable merging feature on AMBA wrapper
Later versions of the AMBA to STBUS convertor introduce the
concept of merging. In essence this allows multiple requests
at the STBUS end to be merged into one AHB burst to the device.

Unfortunately, the PCIe cell will issue a maximal sized read on the
bus, which is not good when talking to the registers on the target
device.

Therefore we have to disable the merging feature which is controlled
by bit 3 of the SD_CONFIG register.

This was first found on Freeman Ultra cut 1.

Signed-off-by: David McKay <david.mckay@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>