stm_pad: Reverse the release order in stm_pad_release
Prior to this change the three resources claimed by the pad manager
(gpio pins, sysconf fields and custom) were released in the same
order they were claimed. However it makes more sense for them to
be freed in the opposite order to the claim. In particular the
custom release needs to come before the sysconf and gpio, as it may
be using either of those resources.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>