stm: stx7100: Correctly free USB gpio's and sysconfigs
In the existing code, stx7100_usb_pad_claim does not free the requested
gpio's and sysconfigs, which creates problem when we build USB as a
module. rmmod was throwing BUG() as the pad manager realizes that gpio
is still in requested state.

Originally two issues were detected while testing hcd_stm driver as
a module.

1. could not do rmmod. which is because of pad manager
   validating the gpio state in pad_release logic.

2. could not insmod after solving problem 1, as sysconf registers were
   not released after claiming in custom pad_claim code.

This patch fixes both the problems, by adding a custom release function
to free the gpio pins, and moving the sysconf claim and initialisation
so it is now handled by the pad manager itself rather than part of the
custom claim. Now the hcd-stm driver can be modprobe/rmmod multiple
times.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>