stm: pm: Added stm_system API
This patch adds the new stm_system API.
It is used to manage the PM chip subsystem (clk, sysconf, gpio, emi)
capability in an ordered priority.

This API has to replace the sysdev approch used until now because
the sysdev design guarantes only that the sysdevice pm-callback
are called when devices are already suspended and the irq are disabled,
but it can not enforce any ordering (priority/hierarchy) on pm callback.

Now under the sysfs entry there will be no more entry as single sysdev device
while there will be only one 'stm_subsystem'

Signed-off-by: Francesco Virlinzi <francesco.virlinzi at st.com>
Reviewed-by: Carmelo Amoroso <carmelo.amoroso at st.com>
Signed-off-by: Chris Smith <chris.smith@st.com>