stm: gpio: hibernation: Restored pin value on direction_out
With this patch during the resume from hibernation
also the pin value is restored for all the pins which
have the direction_out selected.
To avoid a new field the previous 'direction' field
is renamed to 'pm_saved_data' to hold both the
 - pin direction;
 - pin value;

A set of functions allow the management of this field.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi at st.com>
Reviewed-by: Carmelo Amoroso <carmelo.amoroso at st.com>
Signed-off-by: Chris Smith <chris.smith@st.com>