stm_pad: revert pad configuration for legacy stpio interface
The deprecated 'stpio_' interface has no concept of multiple ALT functions.  In
the past, selection of different ALT functions had to be done outside the stpio
API, typically via sys_config registers.  However, since the advent of the pad
manager, sys_config registers that are associated with PIO functions are now
owned by the pad manager.  As a result, it is only safe to use the stpio
interface for GPIO modes of operation.  This approach was taken in the previous
commit 80e0dbfb545013e06cbc85830b6d2b1c1b210b28 "stm_pad: fix regression in GPIO
pad configuration".

However, some software components (e.g. STAPI STPIO) still use the deprecated
stpio interface for ALT mode functions and any required sys_config configuration
is achieved by bypassing the protection offered by the sysconf interface.

Ideally the offending components should be updated to use the pad manager to
request and configure PIO pins.  In the meantime, we revert the stpio pad
configuration to its previous incarnation, and assume the component has some
backdoor to configure the sys_config registers.

This issue was originally discovered while investigating bugzilla #12876.

Signed-off-by: Angus Clark <angus.clark@st.com>