sh_stm: fix SATA/ePCI MiPHY setup for ADI7108 boards (1.0/B/C/D).
This patch fixes the kernel boot on the adi7108 rev D
that hangs due to a broken MiPHY configuration.

This is the status of ePCI and SATA on this board revisions:

-------------------------------------------------------------
 		      |  Ver1.0    VerB   VerC      VerD
SATA 0/1 connector | E-sata0+1 E-sata0 E-sata0 Sata on board
PCI-e Connector    |		   √	   √
-------------------------------------------------------------

Table from:
https://codex.cro.st.com/plugins/docman/?group_id=621&action=show&id=159707

The patch adds a new submenu for this board (like the HDK7108)
and sets the MiPHY for each revisions according to the table above.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Reviewed-by: Francesco Virlinzi <francesco.virlinzi@st.com>