sh_stm: add gmac0 selection for the ADI7108
This patch adds a new option to enable the GMAC0
actually useful for testing. In fact, on this board,
where the RJ45 connector for the GMAC0 can be plugged,
the GMAC0 will be be validated in GiGa mode
(mode never tested on other 7108 boards).

Also the patch removes a warning when compile the kernel:
stmmac0_mdio_bus defined but unused.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Reviewed-by: Nunzio Raciti <nunzio.raciti@st.com>