sh_stm: Enabled NLS support for VFAT in the mb671 defconfig
The mb671_defconfig enables VFAT support but without the proper
configuration of the "iocharset" and the "codepage", so the file
system can't be mounted correctly.

The issue has been discovered by ASV Fast Check and reported in
Bugzilla #13047.

In order to address the issue, the iocharset "iso8859-1" and
codepage "cp437" are enabled in the mb671_defconfig.

Signed-off-by: Francesco Rundo <francesco.rundo@st.com>
Reviewed-by: Carmelo Amoroso <carmelo.amoroso@st.com>