wl12xx: port the platform driver
This patch ports the WL12XX platform driver
useful to build the wl1271 sdio driver that is not
provided into the kernel 2.6.32. The  wl1271 sdio driver
can be built from the compact-wireless sources.
The platform driver is necessary to pass the driver
parameters from the platforms (e.g. IRQ, Voltage etc.).

This driver can be enabled from the "Wireless LAN" menu,
This has been tested on the 7108 only.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>