mmc: debugfs: parse ext_csd via debug_fs
This patch enhances the debug information reported
for the mmc in /sys/kernel/debug/mmc.
It adds a new function to parse the extended CSD registers.
Before this patch, we can only the a big buffer of numbers
that had to be parsed off line according to the JEDEC SPEC.

Signed-off-by: Youssef TRIKI <youssef.triki@st.com>
Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>