Revert "TTY: ldisc, do not close until there are readers"
This reverts commit 9bb7bdfbc6e0699e6daeda46b245ca7964c8bfd4.

This commit added a WARN_ON which has been seen to trigger when using
upstart on a serial console. Despite several comments about this on
LKML its still not clear what the cause is. The change appears to be
completly stand-alone, and so should be safe to revert.

See bugzilla 15076.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>