stm: Fix stx7105 reboot
This patch allows the stx7105 to reboot successfully. The reset chain is
set correctly on boot, and a function, stx7105_prepare_restart, is added
to set the reset period and slow down the EMI clock to allow boot after
a watchdog reset from large SPI flash parts like the M25P128.

Signed-off-by: Chris Smith <chris.smith@st.com>