stmmac: Don't overwrite phy's IRQ if it doesn't support it
The 'probed_phy_irq' field allows a single IRQ number to be passed
in which will be associated with the first available PHY on the bus.
However if a PHY driver has already been bound to the PHY device,
and the PHY driver doesn't support interrupts, this will overwrite
the PHY_POLL flag with an interrupt number which is incorrect.

So add a check to only overwrite the irq if no driver has been bound
to the device or if there is a driver then check that it does indeed
support interrupts.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>