stm: sysconf: Improved flexibility of printing SoC specific sysconf names
Normally sysconf names printed to debugfs consist of the sysconf
group name followed by the offset into the group. For sysconf groups
which don't follow this scheme there is provision for a group specific
function, reg_name. However prior to this change reg_name had to return
a constant string, which limited the flexibility.

This changes reg_name so that it is passed a small buffer into which
the register name must be written. This allows the SoC specific code
to construct the name using sprintf() for example.

As part of this tidy-up the debugfs code in both the pad manager and
the sysconf code are changed to use sysconf_reg_name() for all sysconf
names, and sysconf_group_name() is removed, as it is no longer needed.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>