stm_lirc: fixed wrong RX buffer management.
The RX buffer, used to collect data incoming from the FIFO,
is declared in terms of symbols, consisting of pulse or space.
The RX interrupt handler instead, used the buffer in terms of
pairs (pulse/space) causing in same case access out of buffer boundary.

This patch align all in terms of symbols.
Bugzilla # 16212

Signed-off-by: Angelo Castello <angelo.castello@st.com>
Reviewed-by: Carmelo Amoroso <carmelo.amoroso@st.com>