b2057: Update front panel buttons for new board
The B2057 is using a different version of the front panel (labelled
STI7108-HDK-FP_VerA) which differs from the previous panels in having
a STM8S103 microcontroller, primarily to act as a low power controller.
Currently there is no support for this, and so the extra buttons which
are only attached to this microcontroller cannot be supported, however
the extra buttons which are attached to the TM1668, plus one which is
directly connected to a gpio pin can be supported.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>