sh_stm: initial stxh205 MMC support
This patch adds the initial support for the MMC/SD on the stxh205 boards.

It has been only tested on the B2039 board with the B2048 eMMC modules.

The patch reviews the Kconfig for the B2039 adding the selection of
the B2048 modules. Also it enables the mmc in the defconfig.

At this stage, since the eMMC is a bootable IP the PIO is selected by
the HW at reset no need to re-configure it in SW.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>