libata: Set read/write flag in ata_pio_sector()
The SFF data transfer functions take a parameter to indicate
if we are doing a READ (0) or a WRITE (1) . However, the ata_pio_sector()
function is actually passing the masked ATA flags value through.
So if you are doing a write you will get ATA_TFLAG_WRITE which is
0x8 rather than WRITE (1).

Most driver get away with this because they do something like

if(rw) {
  /* Do something if writing  */
}

However, pata_legacy.c and the sata_stm.c instead do something like

if (rw == WRITE) {
	/* Do something if writing */
}

So will not do the right thing.

Reported-by: Laurent Cans <laurent.cans@sagemcom.com>
Signed-off-by: David McKay <david.mckay@st.com>