libelf: add support for 64 bits ELF parser
Support 64bits ELF parser for those drivers that need to load
64 bits ELF firmware on external microcontrollers.
It needs to select LIBELF_64 option to enable this support.

Use the same pre-processing technique for the libelf.h header
file to define both 32 and 64 version of data types and function
prototypes. This required to create and internal header '_libelf.h'
that could be included twice to create 32 and 64 headers.

Users of libelf keeps including <linux/libelf.h>.

Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>