kptrace: No longer overwrite unconsumed trace
This patch alters how KPTrace behaves when the trace buffer is full.

The relay infrastructure will call subbuf_start_handler when a new
sub-buffer is needed. Previously, if all the sub-buffers were full
(because userspace has not yet consumed them) then unconsumed buffers
would be overwritten. In addition to losing trace data, which can't be
avoided in this case, this caused additional discontinuities in the
trace and made it much harder for userspace to track how much data had
been generated. With this patch, the behavior changes so that if all
sub-buffers are full, no further trace is recorded until one or more
sub-buffers have been consumed. The same amount of data is lost, but in
a more controlled manner that the consumer app and rendering tools can
handle better.

We also warn the user that the buffer is full, and trace is being lost,
with a printk. The solution for them is to request a larger trace buffer
via the consumer app. Because these warnings can be numerous, we never
output more than ten of them for a given trace session.

Signed-off-by: Chris Smith <chris.smith@st.com>