phy: add the EEE support and the way to access to the MMD regs
This patch adds the initial support for the Energy-Efficient
Ethernet (EEE). It has been tested on IC+101G device on ST STB.

To support the EEE we have to access to the MMD registers 3.20 and
3.60/61. So I added two new functions to read/write the MMD
registers (clause 45)/

The upper-layer will invoke the phy_check_eee to properly check
if the EEE is supported by the PHYs.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>