stm: stx7108: Correctly configure TXCLK in RGMII_GTX mode
The TXCLK signal changes polarity in RGMII_GTX mode depending on the
negotiated speed. This requires that the pad retiming is changed at
run time, which previous has not been possible, however with the
introduction of stm_pad_update_gpio() this can now be done.

However to allow this requires storing the struct stm_pad_state so
that the SoC specific code can get access to it, and as the bsp_priv
pointer is already used, this means introducing a new structure to
store the SoC specific data.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>