sh: pmb: extend PMB initialization to better support kexec
The current PMB initialization assumes that the code is running from an
uncached mapping, and so can perform the PMB setting directly.
This assumption is no longer true when the kernel is booted by
kexec tool because the L1 cache is already enabled, and the PMB
setting will have the kernel text mapped cached. However, the uncached
mapping should still remain, so this change adds some checks to see if
there is an existing uncached mapping which can be used directly.

Note that this involves reading the PMB registers from code which is
cached, which is in violation of the ST40 and SH4 documentation. However
this has never been found to be a problem in practice, and conversations
with the architects have confirmed that this should be safe as long as
the code is well isolated (in CPU pipeline terms) from any writes to the
PMB or other MMU related registers.

Note: this will only work in cases where the uncached memory region
of both kernels is covered be the same PMB mapping. Because of the
large size of PMB entries this is frequently the case, but there is
no error recovery when this isn't true.

Signed-off-by: Angelo Castello <angelo.castello@st.com>
[stuart.menefy@st.com: Restructured so cache writeback is an explicit
part of PMB initialization rather then relying on flush in cache_init]
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>